Move theme selection into new panel tab.
