+++ src/modlunky2/ui/levels/custom_levels/options_panel.py
@@ -1,89 +1,8 @@
 import tkinter as tk
 from tkinter import ttk
-from PIL import Image, ImageDraw, ImageEnhance, ImageTk
 
-from modlunky2.mem.state import Theme
 from modlunky2.config import CustomLevelSaveFormat
-from modlunky2.levels.tile_codes import VALID_TILE_CODES
 from modlunky2.ui.levels.custom_levels.save_formats import SaveFormats
-from modlunky2.ui.widgets import PopupWindow, Tab
-
-
-def name_of_theme(theme):
-    if theme == Theme.DWELLING:
-        return "Dwelling"
-    elif theme == Theme.TIDE_POOL:
-        return "Tide Pool"
-    elif theme == Theme.NEO_BABYLON:
-        return "Neo Babylon"
-    elif theme == Theme.JUNGLE:
-        return "Jungle"
-    elif theme == Theme.TEMPLE:
-        return "Temple"
-    elif theme == Theme.SUNKEN_CITY:
-        return "Sunken City"
-    elif theme == Theme.COSMIC_OCEAN:
-        return "Cosmic Ocean"
-    elif theme == Theme.CITY_OF_GOLD:
-        return "City of Gold"
-    elif theme == Theme.DUAT:
-        return "Duat"
-    elif theme == Theme.ABZU:
-        return "Abzu"
-    elif theme == Theme.EGGPLANT_WORLD:
-        return "Eggplant World"
-    elif theme == Theme.ICE_CAVES:
-        return "Ice Caves"
-    elif theme == Theme.OLMEC:
-        return "Olmec"
-    elif theme == Theme.VOLCANA:
-        return "Volcana"
-    elif theme == Theme.TIAMAT:
-        return "Tiamat"
-    elif theme == Theme.HUNDUN:
-        return "Hundun"
-    elif theme == Theme.BASE_CAMP:
-        return "Surface"
-    return "Unknown"
-
-
-def theme_for_name(name):
-    if name == "Dwelling":
-        return Theme.DWELLING
-    elif name == "Jungle":
-        return Theme.JUNGLE
-    elif name == "Volcana":
-        return Theme.VOLCANA
-    elif name == "Olmec":
-        return Theme.OLMEC
-    elif name == "Tide Pool":
-        return Theme.TIDE_POOL
-    elif name == "Temple":
-        return Theme.TEMPLE
-    elif name == "Ice Caves":
-        return Theme.ICE_CAVES
-    elif name == "Neo Babylon":
-        return Theme.NEO_BABYLON
-    elif name == "Sunken City":
-        return Theme.SUNKEN_CITY
-    elif name == "Cosmic Ocean":
-        return Theme.COSMIC_OCEAN
-    elif name == "City of Gold":
-        return Theme.CITY_OF_GOLD
-    elif name == "Duat":
-        return Theme.DUAT
-    elif name == "Abzu":
-        return Theme.ABZU
-    elif name == "Tiamat":
-        return Theme.TIAMAT
-    elif name == "Eggplant World":
-        return Theme.EGGPLANT_WORLD
-    elif name == "Hundun":
-        return Theme.HUNDUN
-    elif name == "Surface":
-        return Theme.BASE_CAMP
-    return None
-
 
 class OptionsPanel(ttk.Frame):
     def __init__(
@@ -91,7 +10,6 @@ class OptionsPanel(ttk.Frame):
         parent,
         modlunky_config,
         zoom_level,
-        on_select_theme,
         on_update_level_size,
         on_update_save_format,
         on_update_hide_grid_lines,
@@ -103,7 +21,6 @@ class OptionsPanel(ttk.Frame):
         super().__init__(parent, *args, **kwargs)
 
         self.modlunky_config = modlunky_config
-        self.on_select_theme = on_select_theme
         self.on_update_level_size = on_update_level_size
         self.on_update_save_format = on_update_save_format
         self.on_update_hide_grid_lines = on_update_hide_grid_lines
@@ -113,74 +30,11 @@ class OptionsPanel(ttk.Frame):
         self.lvl_width = None
         self.lvl_height = None
 
-        self.rowconfigure(2, minsize=20)
-        self.rowconfigure(4, minsize=20)
         self.columnconfigure(0, weight=1)
 
         right_padding = 10
         settings_row = 0
 
-        theme_label = tk.Label(self, text="Level Theme:")
-        theme_label.grid(row=settings_row, column=0, sticky="nsw")
-        self.theme_label = theme_label
-
-        settings_row += 1
-
-        theme_container = ttk.Frame(self)
-        theme_container.grid(row=settings_row, column=0, sticky="nwse")
-        theme_container.columnconfigure(1, weight=1)
-        theme_container.columnconfigure(3, minsize=right_padding)
-
-        # Combobox for selecting the level theme. The theme affects the texture used to
-        # display many tiles and the level background; the suggested tiles in the tile
-        # palette; and the additional vanilla setrooms that are saved into the level
-        # file.
-        self.theme_combobox = ttk.Combobox(theme_container, height=25)
-        self.theme_combobox.grid(row=0, column=0, sticky="nsw")
-        self.theme_combobox["state"] = tk.DISABLED
-        self.theme_combobox["values"] = [
-            "Dwelling",
-            "Jungle",
-            "Volcana",
-            "Olmec",
-            "Tide Pool",
-            "Temple",
-            "Ice Caves",
-            "Neo Babylon",
-            "Sunken City",
-            "Cosmic Ocean",
-            "City of Gold",
-            "Duat",
-            "Abzu",
-            "Tiamat",
-            "Eggplant World",
-            "Hundun",
-            "Surface",
-        ]
-
-        def update_theme():
-            theme_name = str(self.theme_combobox.get())
-            theme = theme_for_name(theme_name)
-            self.theme_select_button["state"] = tk.DISABLED
-            self.theme_label["text"] = "Level Theme: " + theme_name
-            self.on_select_theme(theme)
-
-        self.theme_select_button = tk.Button(
-            theme_container,
-            text="Update Theme",
-            bg="yellow",
-            command=update_theme,
-        )
-        self.theme_select_button["state"] = tk.DISABLED
-        self.theme_select_button.grid(row=0, column=2, sticky="nse")
-
-        def theme_selected(_):
-            self.theme_select_button["state"] = tk.NORMAL
-
-        self.theme_combobox.bind("<<ComboboxSelected>>", theme_selected)
-
-        settings_row += 1
-        settings_row += 1
         # Comboboxes to change the size of the level. If the size is decreased, the
         # tiles in the missing space are still cached and restored if the size is
         # increased again. This cache is cleared if another level is loaded or the
@@ -235,6 +89,7 @@ class OptionsPanel(ttk.Frame):
         self.height_combobox.bind("<<ComboboxSelected>>", size_selected)
 
         settings_row += 1
+        self.rowconfigure(settings_row, minsize=20)
         settings_row += 1
 
         option_header = tk.Label(self, text="Save format:")
@@ -366,19 +221,11 @@ class OptionsPanel(ttk.Frame):
             width=width, height=height
         )
 
-    def update_theme(self, theme):
-        theme_name = name_of_theme(theme)
-        self.theme_combobox.set(theme_name)
-        self.theme_label["text"] = "Level Theme: " + theme_name
-
     def enable_controls(self):
-        self.theme_combobox["state"] = "readonly"
         self.width_combobox["state"] = "readonly"
         self.height_combobox["state"] = "readonly"
 
     def disable_controls(self):
-        self.theme_combobox["state"] = tk.DISABLED
-        self.theme_select_button["state"] = tk.DISABLED
         self.width_combobox["state"] = tk.DISABLED
         self.height_combobox["state"] = tk.DISABLED
         self.size_select_button["state"] = tk.DISABLED
